modified model state management
