+++ client/src/components/battle.tsx
@@ -1,5 +1,3 @@
-import { useEffect, useState } from 'react'
-import axios from 'axios'
 import { createTheme } from '@mui/material'
 import { styled } from '@mui/material/styles'
 import AppBar from '@mui/material/AppBar'
@@ -10,13 +8,15 @@ import Grid from '@mui/material/Grid'
 import LinearProgress from '@mui/material/LinearProgress'
 import Card from '../components/battle/card'
 import { PlayerType, EnemyType, CardType } from '../types/model'
-import { ResPlayer, ResEnemies, ResCards } from '../types/api/response'
 import playerImg from '../images/player.png'
 import enemyImg from '../images/enemy.png'
 import '../styles/battle/style.scss'
 
 type Props = {
   disable: boolean
+  enemies: EnemyType[]
+  player: PlayerType
+  deck: CardType[]
 }
 
 const ENERGY_MAX = 3
@@ -41,133 +41,15 @@ const CustomLinearProgress = styled(LinearProgress)({
 
 
 const Battle = (props: Props): JSX.Element => {
-  const { disable } = props
-  const [enemies, setEnemies] = useState<EnemyType[]>([{
-    id: 0,
-    name: "",
-    imageUrl: "",
-    hp: 0,
-    attack: 0,
-    defense: 0
-  }])
-  const [player, setPlayer] = useState<PlayerType>({
-    name: "",
-    imageUrl: "",
-    hp: 0,
-    attack: 0,
-    defense: 0,
-    energy: 0
-  })
-  const [cards, setCards] = useState<CardType[]>([{
-    name: "",
-    description: "",
-    imageUrl: "",
-    cost: 0,
-    cardType: "",
-    attack: 0,
-    defense: 0
-  }])
-  const [deck, setDeck] = useState<CardType[]>([{
-    name: "",
-    description: "",
-    imageUrl: "",
-    cost: 0,
-    cardType: "",
-    attack: 0,
-    defense: 0
-  }])
-
-  const getEnemies = async (): Promise<void> => {
-    await axios.get(`${process.env.REACT_APP_API_URL_BROWSER}/v1/enemies`)
-    .then(res => {
-      const resEnemies: ResEnemies[] = res.data
-      setEnemies(resEnemies.map((enemy: ResEnemies) => {
-        return {
-          id: enemy.id,
-          name: enemy.name,
-          imageUrl: enemy.image_url,
-          hp: enemy.hp,
-          attack: enemy.attack,
-          defense: enemy.defense
-        }
-      }))
-    })
-    .catch(error => {
-      console.log("敵の取得に失敗しました")
-    })
-  }
-
-  const getPlayer = async (): Promise<void> => {
-    await axios.get(`${process.env.REACT_APP_API_URL_BROWSER}/v1/players`)
-    .then(res => {
-      const resPlayer: ResPlayer = res.data
-      setPlayer({
-        name: resPlayer.name,
-        imageUrl: resPlayer.image_url,
-        hp: resPlayer.hp,
-        attack: resPlayer.attack,
-        defense: resPlayer.defense,
-        energy: resPlayer.energy
-      })
-    })
-    .catch(error => {
-      console.log("プレイヤーの取得に失敗しました")
-    })
-  }
-
-  const getCards = async (): Promise<void> => {
-    await axios.get(`${process.env.REACT_APP_API_URL_BROWSER}/v1/cards`)
-    .then(res => {
-      const resCards: ResCards[] = res.data
-      const newCards: CardType[] = resCards.map((card: ResCards) => {
-        return {
-          name: card.name,
-          description: card.description,
-          imageUrl: card.image_url,
-          cost: card.cost,
-          cardType: card.card_type,
-          attack: card.attack,
-          defense: card.defense
-        }
-      })
-      setCards(newCards)
-      initializeDeck(newCards)
-    })
-    .catch(error => {
-      console.log("カードの取得に失敗しました")
-    })
-  }
+  const { disable, enemies, player, deck } = props
 
   const nameplate = (): JSX.Element[] => {
-    return cards.map((card, index) =>
+    return deck.map((card, index) =>
       <Grid item xs={2} key={index}>
         <Card card={card} />
       </Grid>
     )
   }
-
-  const initializeDeck = (cardList: CardType[]): void => {
-    const defaultDeck: CardType[] = []
-    const strike: CardType | undefined = cardList.find((card: CardType) => card.name === "ストライク")
-    const protection: CardType | undefined = cardList.find((card: CardType) => card.name === "ぼうぎょ")
-    if (strike !== undefined) {
-      for (let i = 0; i < 5; i++) {
-        defaultDeck.push(strike)
-      }
-    }
-    if (protection !== undefined) {
-      for (let i = 0; i < 5; i++) {
-        defaultDeck.push(protection)
-      }
-    }
-    setDeck(defaultDeck)
-  }
-
-  useEffect((): void => {
-    getEnemies()
-    getPlayer()
-    getCards()
-  }, [])
 
   return (
     <div style={{ display: disable ? 'none' : '' }}>
